refactor: move useActiveSection hook to separated file
